Implement h=0 or w=0 for thumbnail and fit ops.
The usage section of the documentation says:

width - The desired width of the image, if 0 is provided the service will
calculate the ratio with height
height - The desired height of the image, if 0 is provided the service will
calculate the ratio with width

But picfit panics if you attempt this with any op other than resize.  This
change makes it work with other ops.